+++ CHANGELOG.md
@@ -156,6 +156,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added nix::fcntl::{rename, renameat}.
   ([#752](https://github.com/nix-rust/nix/pull/752))
 
+- Added `nix::unistd::unlinkat`.
+  ([#753](https://github.com/nix-rust/nix/pull/753))
 
 ### Changed
 - Exposed the `mqueue` module for all supported operating systems.